accessible/tests/mochitest/tree/test_txtctrl.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>

<window xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        title="Accessible XUL textbox and textarea hierarchy tests">

  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js" />

  <script type="application/javascript"
          src="../common.js" />
  <script type="application/javascript"
          src="../role.js" />
  <script type="application/javascript"
          src="../events.js" />

  <script type="application/javascript">
  <![CDATA[
    //gA11yEventDumpToConsole = true;
    //enableLogging("tree,verbose"); // debug stuff

    ////////////////////////////////////////////////////////////////////////////
    // Test

    function doTest()
    {
      //////////////////////////////////////////////////////////////////////////
      // textboxes

      var accTree =
        { SECTION: [
          { ENTRY: [ { TEXT_LEAF: [] } ] },
          { MENUPOPUP: [] }
        ] };

      // default textbox
      testAccessibleTree("txc", accTree);

      //////////////////////////////////////////////////////////////////////////
      // search textbox
      accTree =
         { GROUPING: [
           { ENTRY: [ { TEXT_LEAF: [] } ] },
         ] };
      testAccessibleTree("txc_search", accTree);

      //////////////////////////////////////////////////////////////////////////
      // search textbox with search button

      if (MAC) {
         accTree =
           { GROUPING: [
             { ENTRY: [ { TEXT_LEAF: [] } ] },
           ] };
      } else {
         accTree =
           { GROUPING: [
             { ENTRY: [ { TEXT_LEAF: [] } ] },
             { PUSHBUTTON: [] },
           ] };
      }

      testAccessibleTree("txc_search_searchbutton", accTree);

      //////////////////////////////////////////////////////////////////////////
      // number textbox

      accTree =
        { SECTION: [
          { SPINBUTTON: [
            { ENTRY: [ { TEXT_LEAF: [] } ] },
            { PUSHBUTTON: [ ] },
            { PUSHBUTTON: [ ] }
          ] },
          { MENUPOPUP: [] },
        ] };

      testAccessibleTree("txc_number", accTree);

      //////////////////////////////////////////////////////////////////////////
      // password textbox

      accTree =
        { SECTION: [
          { PASSWORD_TEXT: [ { TEXT_LEAF: [] } ] },
          { MENUPOPUP: [] }
        ] };

      testAccessibleTree("txc_password", accTree);

      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  ]]>
  </script>

  <hbox flex="1" style="overflow: auto;">
    <body xmlns="http://www.w3.org/1999/xhtml">
      <a target="_blank"
         href="https://bugzilla.mozilla.org/show_bug.cgi?id=542824"
         title="Create child accessibles for text controls from native anonymous content">
        Mozilla Bug 542824
      </a><br/>
      <p id="display"></p>
      <div id="content" style="display: none">
      </div>
      <pre id="test">
      </pre>
    </body>

    <vbox flex="1">
      <textbox id="txc" value="hello"/>
      <box role="group" id="txc_search">
        <search-textbox value="hello"/>
      </box>
      <box role="group" id="txc_search_searchbutton">
        <search-textbox searchbutton="true" value="hello"/>
      </box>
      <textbox id="txc_number" type="number" value="44"/>
      <textbox id="txc_password" type="password" value="hello"/>
    </vbox>
  </hbox>

</window>
